requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
ffmpeg-python
transformers
sentencepiece
sacremoses
accelerate